# UAS/src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "This Is API Documentation for My Application",
        "termsOfService": "https://pemweb.uas/terms",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://pemweb.uas",
            "description": "Server Lokal"
        },
        {
            "url": "https://pemweb.uas",
            "description": "Local API Server"
        }
    ],
    "paths": {
        "/api/upload": {
            "post": {
                "tags": [
                    "Files"
                ],
                "summary": "Upload a file",
                "operationId": "b4116c9c1cbcf1030ec056ee6176a3dc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file",
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the file",
                                        "type": "string"
                                    },
                                    "file": {
                                        "description": "File to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "File uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "File uploaded successfully"
                                        },
                                        "filename": {
                                            "type": "string",
                                            "example": "1721531741_document.pdf"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "application/pdf"
                                        },
                                        "size": {
                                            "type": "integer",
                                            "example": 204800
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Invoice July"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid upload request"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/files": {
            "get": {
                "tags": [
                    "Files"
                ],
                "summary": "List all uploaded files",
                "operationId": "9f13ba9d8989298d8ac2a6a67c7984d6",
                "responses": {
                    "200": {
                        "description": "List of files",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/File"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/update/{id}": {
            "post": {
                "tags": [
                    "Files"
                ],
                "summary": "Update a file by ID",
                "operationId": "6b2176bd5bc61ba03efd05802756cdd1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "File updated successfully"
                    },
                    "404": {
                        "description": "File not found"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/upload/{id}": {
            "delete": {
                "tags": [
                    "Files"
                ],
                "summary": "Delete a file by ID",
                "operationId": "65bab27e932e3995c86bef12267e53dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File deleted successfully"
                    },
                    "404": {
                        "description": "File not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "File": {
                "title": "File",
                "description": "Schema untuk data file",
                "required": [
                    "id",
                    "name",
                    "path",
                    "type",
                    "size"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "My File.pdf"
                    },
                    "path": {
                        "type": "string",
                        "example": "files/1721531741_document.pdf"
                    },
                    "type": {
                        "type": "string",
                        "example": "application/pdf"
                    },
                    "size": {
                        "type": "integer",
                        "example": 204800
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-21T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-21T10:01:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Files",
            "description": "Files"
        }
    ]
}